requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.4.0
scikit-image>=0.18.0
tensorflow>=2.8.0
torch>=1.11.0
torchvision>=0.12.0
rasterio>=1.2.0
scikit-learn>=1.0.0
seaborn>=0.11.0
tqdm>=4.62.0
albumentations>=1.3.0